refactor(daemon): extract media task registry into media/task-registry.ts
Strangler-fig slice 2 of the server.ts decomposition (follows the
startChatRun extraction in PR #5128). Moves the in-memory media-task
registry — the `mediaTasks` Map plus hydrate/create/persist/append/
notify/snapshot helpers and the TTL constant — out of server.ts into a
new `media/task-registry.ts` sibling module, layered over the existing
SQLite persistence in `media/tasks.ts`.

Byte-identical move: function bodies are unchanged; server.ts imports the
helpers back and wires them into startServer's boot rehydration and the
media route deps object exactly as before. The four db helpers the moved
code needs (get/insert/update/deleteMediaTask) now import into the new
module and leave server.ts's import block; the three still used by
startServer (listMediaTasksByProject, listRecentMediaTasks,
reconcileMediaTasksOnBoot) stay.

server.ts 6172 -> 6062 lines. Public/runtime behavior unchanged.

Validation: tsc -p apps/daemon/tsconfig.json --noEmit clean; pnpm guard
clean; chat-run-sse-shapes characterization 4/4; media tasks-persistence
+ tasks-routes suites green (14/14); policy-routes' 7 failures proven
pre-existing (identical on base d20acc87).
